multi-package: extract the shared logic to a separate package
To make sharing the common/agnostic logic between different CI system
implementations, we will extract the shared code into a library package
found under 'shared/', and move the GitHub Action code into the
'github-action/' directory.

Update the references to reflect the new workspaces, as well as the
various config files for TS etc.

In the new root package.json we add convienence scripts to run the usual
`build` or `test` on all workspaces.

Signed-off-by: Matthew John Cheetham <mjcheetham@outlook.com>